# server/src/swagger.yaml
openapi: 3.0.3
info:
  title: MeRN
  description: |
    API description for MeRightNow social network. 

    All endpoints need authentication, so first, create an user or login with one.
  version: 1.0.0
  contact: {}
servers:
  - url: 'https://api-merightnow.onrender.com/mern/v1'
  - url: 'http://localhost:8000/mern/v1'
paths:
...
  /users/{id}/publications:
    get:
      tags:
        - Users
      summary: Get all publications by user
      description: Get all publications by user
      operationId: getAllPublicationsByUser
      parameters:
        - name: id
          in: path
          description: User's id. Can be mongoId or username.
          required: true
          schema:
            type: string
            example: user2
        - name: after
          in: query
          description: Minimum date to be searched. If omitted, there is no lower limit in the search.
          schema:
            type: string
            example: '1999-04-22T12:58:03.378Z'
        - name: before
          in: query
          description: Maximum date to be searched. If omitted, there is no upper limit in the search.
          schema:
            type: string
            example: '2023-04-22T12:58:03.378Z'
        - name: sort
          in: query
          description: |
            Valid attributes are: createdAt,updatedAt.

            The sorting can be performed on several attributes by concatenating them with the character ",".

            The default sorting order is ascending, indicate the "-" char in front of the attribute to be sorted descending.

            Default sort: -updatedAt.
          schema:
            type: string
            example: '-updatedAt'
        - name: page
          in: query
          description: Number of page.
          schema:
            type: string
            example: '1'
      responses:
        '':
          description: ''
...
  /publications:
    get:
      tags:
        - Publications
      summary: Get all publications
      description: Get all publications
      operationId: getAllPublications
      parameters:
        - name: after
          in: query
          description: Minimum date to be searched. If omitted, there is no lower limit in the search.
          schema:
            type: string
            example: '1999-04-22T12:58:03.378Z'
        - name: before
          in: query
          description: Maximum date to be searched. If omitted, there is no upper limit in the search.
          schema:
            type: string
            example: '2023-04-22T12:58:03.378Z'
        - name: sort
          in: query
          description: |
            Valid attributes are: createdAt,updatedAt.

            The sorting can be performed on several attributes by concatenating them with the character ",".

            The default sorting order is ascending, indicate the "-" char in front of the attribute to be sorted descending.

            Default sort: -updatedAt.
          schema:
            type: string
            example: '-updatedAt'
        - name: q
          in: query
          description: Username, title or description to search.
          schema:
            type: string
            example: ''
        - name: page
          in: query
          description: Number of page.
          schema:
            type: string
            example: '1'
      responses:
        '':
          description: ''
    post:
      tags:
        - Publications
      summary: Create publication
      description: Create publication
      operationId: createPublication
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                description:
                  type: string
                  description: Publication's description.
                  example: ''
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Publication's images (files).
                title:
                  type: string
                  description: Publication's title.
                  example: testpub
      responses:
        '':
          description: ''
  /publications/{id}:
    get:
      tags:
        - Publications
      summary: Get publication by id
      description: Get publication by id
      operationId: getPublicationById
      parameters:
        - name: id
          in: path
          description: Publication's id.
          required: true
          schema:
            type: string
            example: 63e6608d2397e6b0a1e0f9a4
      responses:
        '':
          description: ''
    delete:
      tags:
        - Publications
      summary: Delete publication
      description: Delete publication
      operationId: deletePublication
      parameters:
        - name: id
          in: path
          description: Publication's id.
          required: true
          schema:
            type: string
            example: 63e660983b381b1a6ba422f5
      responses:
        '':
          description: ''
    patch:
      tags:
        - Publications
      summary: Update publication
      description: Update publication
      operationId: updatePublication
      parameters:
        - name: id
          in: path
          description: Publication's id.
          required: true
          schema:
            type: string
            example: 63e66070e7b05efa94ac218d
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                description:
                  type: string
                  description: Publication's description.
                  example: descriptionupdated
                newImg:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Publication's new images (files).
                oldImg:
                  type: array
                  items:
                    type: string
                  description: Publication's old images (old urls).
                title:
                  type: string
                  description: Publication's title.
                  example: titleupdated
      responses:
        '':
          description: ''
...
  /publications/{id}/like:
    patch:
      tags:
        - Publications
      summary: Like publication
      description: Like publication
      operationId: likePublication
      parameters:
        - name: id
          in: path
          description: Publication's id.
          required: true
          schema:
            type: string
            example: 63e66070e7b05efa94ac218d
      responses:
        '':
          description: ''
  /publications/{id}/dislike:
    patch:
      tags:
        - Publications
      summary: Dislike publication
      description: Dislike publication
      operationId: dislikePublication
      parameters:
        - name: id
          in: path
          description: Publication's id.
          required: true
          schema:
            type: string
            example: 63e66070e7b05efa94ac218d
      responses:
        '':
          description: ''
...